Add ability to whitelist specific functions to the `AbstractFunctionRestrictions` class.
This is only useful when used in combination with blocking a function group using a wildcard in the function name.

However, it does make it more forward compatible. This allows to still block a complete (PHP) extension , even for new functions being added to the extension, while whitelisting - for instance - a WP native function which was named a bit unfortunately.