Fix VIP restricted function error.
The `urlencode` check was checking for `rawurlencode()` instead of `urlencode()` while at the same time advising to use `rawurlencode()`.
The grammer of the error message was incorrect as well.
Both are now fixed.

In the [same commit](e7707b2338e7823a2d14cfbdca8bdc88488d6081) as this was introduced, a number of other restricted functions were added, but none of them were accompanied by unit tests.
This is now fixed as well.