Defer to upstream eval() sniff.
`eval()` is a language construct, not a function.

To detect it, the `T_EVAL` token was included in the AbstractFunctionRestrictionsSniff, but this token was only used for detecting the use of `eval()`, not for further interpretation of what was being eval-ed.
To detect `eval()` in that way is highly inefficient as
a) all function related sniffs using the abstract now also (unnecessarily) processed the `T_EVAL` token
b) the logic contained in the abstract would be unnecessarily executed for a `T_EVAL` token.

This PR makes the following changes:
* No longer include the `T_EVAL` token in the tokens being sniffed for by the AbstractFunctionRestrictionsSniff.
* Defer to an existing upstream sniff to detect the use of `eval()`. Adjust the message type and content to be in line with what has been used up to now.
* Remove the `WordPress.PHP.RestrictedFunctions` sniff as it only contained the check for `eval()`.
* Update the relevant documentation:
    - Remove references to `eval` from the `getGroups()` method documentation
    - Update the changelog in the `WordPress.VIP.RestrictedFunctions` sniff