:sparkles: New CapitalPDangitSniff
This PR adds a new sniff which checks whether `WordPress` is spelled correctly in text strings, comment text and class/interface/trait names and can catch a wide range of typical misspellings.

For this sniff, I've erred on the side of caution as this is **_very_** prone to false positives.
This means that some misspellings will likely be missed by this sniff, but that seemed the better choice than to have everyone be terribly annoyed by the sniff being really buggy.

Notes:
        * The sniff has been added to the `WordPress-Core` ruleset.
        * The sniff will throw a `warning` when a misspelling is found.
        * The sniff will only throw one warning per token, even when more than one misspelling is found in the same phrase. The warning will in that case contain a count of the number of misspellings found + list them.
        * For misspellings in text strings and comments, the sniff includes a fixer. For class names and the like, no fixer is included as that will always need attention from a developer.
        * The fixer has been primed to be able to deal with text strings which contain exceptions as well as misspellings which should be corrected, and will deal with this properly.
        * The sniff adds a new whitelist comment `spelling`, though this can obviously not be used to whitelist misspellings in comments.

The sniff takes the following situation into account and will ignore these when encountered:
        - URLs for wordpress.org/com/net/tv.
        - `@...` usernames starting with `wordpress`
        - email addresses with a domain starting with `wordpress`
        - email addresses with a user name ending with `wordpress` (typical default email address set up by core)
        - (most) variable names (as used in docblock @param comments).
        - directory paths containing a folder starting or ending with `wordpress`.
        - file names containing `wordpress` for a limited set of extensions.
        - `wordpress` prefixed or suffixed with dashes as those are indicators that the
           term is probably used as part of a CSS class, such as `fa-wordpress`
           or filename/path like `class-wordpress-importer.php`.
        - back-tick quoted `wordpress` (to help with deliberate misspellings in comments)
        - URL query strings containing `wordpress`
        - Any docblock `@link` tag content (if not previously ignored already)
        - HTML attributes containing `wordpress`
        - escape quoted `\`wordpress\`` as used in xpath queries.

The sniff will ignore *anything* within an array, as well as string array keys like `$var['wordpress']` as these will generate false positives in 80% of the cases.
The sniff will skip forward to the end of the array when one is encountered. This also prevents the sniff "hanging" when processing large configuration arrays.

The sniff is accompanied by extensive unit tests and has been run over both WP core as well as some 180+ recently updated themes (7500+ php files) to try and eliminate most false positives.

The only "common" (1 instance in core) false positive I've not eliminated is when the term is used in a parameter in a function call. Excluding that would also exclude examining text string passed to translation functions which would be counter-productive.