PHPCS 3.x compat: Add helper class & implement use
A number of PHPCS 2.x classes have been split out over several classes in PHPCS 3.x.
Aliasing the new classname to the old would therefore not be a good idea.

This helper class acts as a compatibility layer for these cases.
The helper class only addresses the most common functionality needed and is by no means exhaustive.
The helper class contains one function which is currently not (yet) used `get_version()`. This method was added in anticipation of continuing support for both PHPCS 3.x as well as 2.x for another year. PHPCS 3.x is likely to bring functionality which will not be available in PHPCS 2.x, so being able to retrieve the PHPCS version will, at some point, probably be helpful.

N.B. The class name for this class has to be in CamelCaps (without underscore) as otherwise the PHPCS 2.x autoloader can not find it.