:sparkles: New `WhiteSpace.PrecisionAlignment` sniff for `Core`
This sniff complements the upstream `Generic.WhiteSpace.DisallowSpaceIndent` sniff.

The `Generic.WhiteSpace.DisallowSpaceIndent` sniff makes sure that tabs are used for indentation instead of spaces.
However, the sniff allows for precision alignment of `# spaces < tab_width`.

In WP it is not customary to use precision alignment, though there are a few exceptions where it is used intentionally.

This means that precision alignment can not be auto-fixed.

We can, however, throw a warning when precision alignment is detected which is what this sniff implements.

Notes:
* This sniff will detect precision alignment in PHP, JS as well as CSS files.
* The sniff adds a new `precision alignment` whitelist comment.
* The sniff will handle the "space before star"precision alignment for comments without issue.
* The sniff adds a new custom property which allows users to pass an array of tokens to ignore for the purposes of this sniff. By default nothing is ignored.

Includes unit tests.

I've added the sniff to the `Core` ruleset as this is one of the issues which came out of the initial review of auto-fixed core files.

A run of the sniff over core (using the current `phpcs.xml.5.dist` basis) results in 281 infractions in 68 files. See the attachment for details.
A visual inspection shows no false positives and that nearly all of the infractions found should be corrected. There are only very few were by the looks of it, the precision alignment was actually intended as such.

Fixes 983