ArrayDeclarationSpacing: add a property to make an exception for single item associative arrays
Allow for a property to be set in a custom ruleset to allow single item associative arrays to be single line.

If someone can come up with a better (less lengthy) property name: please let me know and I'd happily adjust the PR.

I've chosen to implement this as a `true`/`false` toggle.
The toggle defaults to `false` based on the rule as it is in the handbook.

Alternatively, it could be implemented numerically, i.e. number of items needed before the associative array rule take effect with a default of `0`, but that felt like it would make the rule very arbitrary where one project would have it at `1`, another at `3` etc., which would counter the use of having the rule at all.

Fixes 1104

----

- [ ] Add info on the new property to the customizable properties wiki page