:sparkles: New "Generic.PHP.DiscourageGoto" sniff
Usage of goto is seen as a typical code smell.

Includes unit tests.

Adding this sniff to `Extra` for now. The sniff can be moved to `Core` once the handbook has been adjusted (if it will be).

This sniff is a duplicate of the same as pulled upstream in PR squizlabs/PHP_CodeSniffer 1664
Once the minimum PHPCS requirement for WPCS goes up beyond the version in which the upstream PR  is merged, this sniff can be safely removed.