ArrayIndentation: deal correctly with trailing comments
If a previous array item has a trailing comment, the sniff would bow out as it would defer to the ArrayDeclarationSpacing sniff to fix a multi-line array where items are not all on their own line first.
However, in the case of a trailing comment and the next array item in actual fact starting on its own line, the ArrayDeclarationSpacing sniff would (correctly) not kick in and the indentation error for the item would never be addressed.

This fixes that. Includes unit tests.

N.B.: Verified & confirmed that the ArrayDeclaration sniff already handled this situation correctly.

Fixes 1179