:sparkles: Add new WhiteSpace.SemicolonSpacing sniff
This new sniff extends the upstream `Squiz.WhiteSpace.SemicolonSpacing` sniff but allows for semi-colons used to denote an empty condition within a `for()` control structure to have a space before them.

Once this has been fixed upstream and the WPCS minimum required PHPCS version has gone up beyond the PHPCS version containing the fix, this custom sniff can be removed and we can defer to the upstream sniff again.

Refs:
* http://php.net/manual/en/control-structures.for.php
* squizlabs/PHP_CodeSniffer 1690