I18n: improve handling of "default" text-domain
PR 1141 changed the behaviour of the I18n sniff to ignore a missing text domain argument if `default` was one of the allowed text-domains.

This was overshooting a little as this effectively disabled the complete "missing domain argument" check if `default` was one of the passed text-domains.

This PR changes that behaviour as well as addressing the remaining comment made in the original PR:
> If the text_domain in the ruleset is `default` and the `$text_domain` has explicitly been set to default as well, then this would seem like a good redundancy warning to trigger, though it shouldn't be an outright error.
>
> I would like to suggest making it a fixableWarning in that case.

Ref: 1141 #issuecomment-329317824

If `default` is _the only text-domain_ set in the ruleset:
* The sniff will now *not* trigger an error for a missing text-domain (= same as in PR 1141).
* Trigger an auto-fixable warning about the text-domain argument being redundant when the text-domain in a gettext function call is explicitly set to `default`.

If `default` is _one of an array of text-domains_ set in the ruleset:
* The sniff will now trigger a warning instead of an error for a missing text-domain. The warning has a slightly different error message and a different error code. This fixes the "overshooting".

Additionally, this PR also:
* Fixes a potential "undefined index" notice for line 375.
* Reduces some code duplication in the examination of text string tokens.

Includes adjusted unit tests.

Fixes 1212