PreparedSQLPlaceholders: update for WP 4.8.3 changes
WP 4.8.3 brings back numbered and formatted placeholders. The sniff was originally written to detect *against* those, so to prevent confusing devs terribly, something had to be done quickly.

So here goes.

* Allow for normal `vsprintf()` placeholders, but only for digits, floats and strings - `dfFs`.
* Continue to detect unsupported placeholders, though this has now become a far more complex set to detect.
* Continue to error on quoted _simple_ placeholders.
* New: Throw a warning on _unquoted_ **complex** placeholders. These will not be quoted by `$wpdb->prepare()` as these are often used to pass in partial query strings.
* Downgrade the mismatched placeholders vs replacements message to a `warning`. As the passing in of _partial query strings_ has been brought back, the chances of a number mismatch here have increased, so a warning seems more appropriate.

The regexes are "variations on the same theme" as the regex which is used in the `WP.I18n` sniff, but with some significant differences.

The first regex is not that different, except that it better protects against issues with a literal `%` (`%%`) being followed by a placeholder. It is also limited to the allowed placeholder types.
The second regex has more significant differences, both in the greediness as well as the matching of "whatever comes after" the `%`.

Includes updated & adjusted unit tests to test for the new reality.

This should be the last PR for WPCS `0.14.0`.