Readme: move IDE setup instructions to the wiki
I've created wiki pages containing the information which is (was) in the readme about configuring Sublime Text and Atom for use with PHPCS & WPCS.

This PR removes that information from the Readme and adds links to the wiki pages in it's place.

It also simplifies the IDE list to a bullet list.