Sniff: support combining PHPCS annotations with WPCS whitelist comments
The new PHPCS 3.2.0+ annotations are line precise, while the WPCS whitelist comments are statement precise.

This change allows combining the two for the time being, by using the `-- for reasons` part of the PHPCS annotation to add a WPCS whitelist comment.

Includes unit test in the EscapeOutput unit test case file.

Note: the sniff name in the `phpcs:` annotation in the unit test case is mangled on purpose. This way we can test that the error is not suppressed by PHPCS because of the annotation, but is ignored because of the WPCS whitelist comment.