Build: move WPCS's own ruleset to the root directory
This file was initially placed in the `bin` directory so as not to confuse users looking for the example custom ruleset.

By renaming the file to `.phpcs.xml.dist` - take note of the `.` -, which is possible since PHPCS 3.1, the file is no longer listed directly next to the example ruleset, so it should be fine to move it to the repo root directory.