Deprecation messages: Simplify & improve code
The `addWarning()` and `addError()` methods return a boolean value signifying whether or not the notice has been thrown. This will in most cases be `true`, but will return `false` when the notice for whatever reason is being ignored.

Instead of hard-setting the `$this->thrown` value to true, let's use the return value instead so if the message is ignored for one file, it will still be thrown for the next.