AlternativeFunctions: improve handling of strip_tags() and parse_url()
The `wp_strip_all_tags()` function does not allow for passing a list of allowed tags, while the PHP native function does, so the WP alternative is not a valid alternative when the second parameter `$allowable_tags` is passed.

Along the same lines, the `wp_parse_url()` function only added support for the PHP native second parameter `$component` in WP 4.7.0.

The changes now made means the sniff will bow out from throwing an error if either of the above situations is detected.

Includes unit tests.