Abstracts: change $exclude property from string to array
The `$exclude` property which is included in most abstract sniffs and by extension in all sniffs which use these abstracts, used to expect a comma-delimited list.
This comma-delimited list was then exploded to an array from within the sniff.

In other words, in reality the sniff expected an `array`.

This has now been formalized. All `$exclude` properties now expect `array` input.

As outlined in 1368, this is not a BC-break as properties passed as a comma-delimited string will still be supported (for now). Support for this may be removed at some point in the future.

The upside of making this change is that users who use WPCS 3.3.0, can start using the new property array format in their custom rulesets which makes for a more easily readable ruleset.

- [ ] Update the wiki section about the `exclude` property in custom rulesets.

Fixes 1368