Travis: add builds to test against minimum supported PHPCS 3.x version
Since PHPCS 3 came out a number of minor versions have been released.
Currently, the builds were only tested for the PHPCS 3.x range against `master` ( currently `3.3.1-alpha`), while PHPCS 3.0.2 is the minimum supported PHPCS 3.x version.

As there have been numerous changes to the tokenizer between 3.0.2 and `master` which could influence the accuracy of the sniff results, the tests should be run against the minimum supported PHPCS 3.x version too.

Unfortunately PHPCS 3.0.2 still contained a bug in the new autoloader which caused a fatal error when unit tests would be run for sniffs in external standards extending other sniffs in the same external standard, so the actual minimum PHPCS 3.x version on which the unit tests can be run is `3.1.0`.