New WPCS native `NoSilencedErrors` sniff
Until now, WPCS used the PHPCS native `Generic.PHP.NoSilencedErrors` sniff. This sniff would throw warnings for all uses of the `@` error silencing operator. However, there are a number of PHP functions for which the use of the error silencing operator is warranted as no amount of error checking can prevent warnings from being thrown.

This PR adds a new WPCS native `PHP.NoSilencedErrors` sniff which will:
* not throw warnings when the error silencing operator is used in combination with one of the whitelisted functions.
* throw warnings in all other cases.
* allows for a user-defined list of additional whitelisted functions to be passed via a custom ruleset.
* displays a brief snippet of code to show the context in which the error silencing operator is being used.

To get to the current list of whitelisted functions, I've checked:
* The PHP manual for every single function in the [filesystem](http://php.net/manual/en/ref.filesystem.php), [directory](http://php.net/manual/en/ref.dir.php),  [stream](http://php.net/manual/en/ref.stream.php), [ftp](http://php.net/manual/en/ref.ftp.php), [gz](http://php.net/manual/en/ref.zlib.php) and [ssh2](http://php.net/manual/en/ref.ssh2.php) extensions.
* Ran the sniff over WP Core and used the reported metrics to verify a number of additional functions for which WP Core currently uses error silencing.

Functions for which there is a way to avoid the warning being thrown have **not** been added to the list. So, for example, the `header()` function does not qualify as the warning could be prevented by wrapping the function call in a `headers_send()` condition.

Fixes 1449