Core ruleset: change behaviour for the `PEAR/FunctionCallSignature` sniff
The `allowMultipleArguments` property of the `PEAR.Functions.FunctionCallSignature` sniff defaults to `true`.

By setting it to `false`, the sniff behaviour changes and demands that there is just one parameter on each line in a multi-line function call.
The behaviour for single-line function calls is not affected by this change.

This property has been added to the WP Core native custom ruleset in [changeset 43571](https://core.trac.wordpress.org/changeset/43571) as part of [trac ticket 44600](https://core.trac.wordpress.org/ticket/44600).

This PR now formalizes this by adding the property to the `WordPress-Core` ruleset.

Note: The WP Core handbook still needs to be adjusted to mention this change!

While this is a significant change, IMHO it is not a *breaking* change, but an enhancement, i.e. something on which WPCS previously had no opinion, will now start throwing errors.
With that in mind, I'm targeting this PR for WPCS 1.1.0.

Regarding the removal of the severity changes:
Previously these would already kick in when using the fixer, but would *not* throw an error when just checking the code with `phpcs`.

The impact of these changes is as follows:
Original code:
```php
$a = myfunction( $param1,
    $param2, $param3 );
```

... would previously have been auto-fixed as:
```php
$a = myfunction(
	$param1,
	$param2, $param3
);
```

Code layout which would still be allowed when just the property change would be applied:
```php
$a = myfunction( $param1,
    $param2,
    $param3 );
```

Code layout expected and auto-fixed as, because of the additional removal of the severity changes:
```php
$a = myfunction(
	$param1,
	$param2,
	$param3
);
```