PHPCS 3.2.3+: Remove the Sniff::$phpcsCommentTokens property
The (protected) `Sniff::$phpcsCommentTokens` property was a work-around for the `Tokens::$phpcsCommentTokens` array not being available prior to PHPCS 3.2.3.

This work-around is no longer needed now the minimum PHPCS version is 3.3.1.

Note:
This is a BC-break for external standards extending the WordPressCS `Sniff` class which may be using this property. They should adjust their code to use the PHPCS native `Tokens::$phpcsCommentTokens` array instead.